+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 
 * Add files_changed
 * Tables for display
+* Add state to pull requests table
 
 
 0.1.2 (2017-07-05)